pyproject.toml
[project]
name = "lazysoc"
version = "0.1.0"
description = "SoC 寄存器生成工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.scripts]
lazysoc = "lazysoc.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
strict = true
python_version = "3.10"

[tool.hatch.build.targets.wheel]
packages = ["src/lazysoc"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
